[REFACTORING] Split IMAPServerAdminUsersTest
